+++ CHANGES.rst
@@ -23,6 +23,10 @@ upgrading your version of coverage.py.
 Unreleased
 ----------
 
+- The ``debug data`` command now takes file names as arguments on the command
+  line, so you can inspect specific data files without needing to set
+  ``COVERAGE_FILE``.
+
 - Fix: coverage.py uses a more disciplined approach to detecting where
   third-party code is installed, and avoids measuring it. This shouldn't change
   any behavior. If you find that it does, please get in touch.